% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_relatorio_dez}
\alias{download_relatorio_dez}
\title{Download de relatorio de emplacamento de dezembro}
\usage{
download_relatorio_dez(ano)
}
\arguments{
\item{ano}{Ano do relatorio}
}
\description{
Download de relatorio de emplacamento de dezembro
}
\examples{
download_relatorio_dez(2024)
}
