% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emplacamento_anual}
\alias{emplacamento_anual}
\title{Quantidade de veiculos emplacados, por ano e modelo}
\format{
Uma tibble com 2200 linhas and 5 variáveis:
\describe{
  \item{tipo}{Se automovel ou utilitário}
  \item{montadora}{Montadora que produz o modelo}
  \item{modelo}{Nome do modelo}
  \item{quantidade}{Quantidade de veiculos emplacados no ano}
  \item{ano}{Ano do emplacamento}
}
}
\usage{
emplacamento_anual
}
\description{
Um conjunto de dados que contem os 50 principais modelos de automoveis e
 comerciais leves de cada ano (2003-2024).
 As variaveis são as seguintes:
}
\keyword{datasets}
