% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{leitura_relatorio_dez}
\alias{leitura_relatorio_dez}
\title{Leitura de relatorio de dezembro}
\usage{
leitura_relatorio_dez(ano)
}
\arguments{
\item{ano}{Ano do relatorio}
}
\value{
Uma tibble com os dados do acumulado do ano
}
\description{
Leitura de relatorio de dezembro
}
