% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{link_relatorio_dez}
\alias{link_relatorio_dez}
\title{Download de relatorio de emplacamento de dezembro}
\usage{
link_relatorio_dez(ano)
}
\arguments{
\item{ano}{Ano do relatorio}
}
\value{
O link para o relatorio
}
\description{
Download de relatorio de emplacamento de dezembro
}
